+++ tvsd/search.py
@@ -5,20 +5,24 @@ import pyclbr
 import sys
 from typing import TYPE_CHECKING, Any, Literal, Union, List
 from bs4 import PageElement
+from tvsd.sources import *
+from rich.table import Table
+from rich.console import Console
+
 
 import typer
 
 
 from tvsd import sources
 from tvsd.source import Source
-from tvsd.sources.xiao_bao import XiaoBao
-from tvsd.sources.olevod import OLEVOD
 
 from tvsd.utils import LOGGER
 
 if TYPE_CHECKING:
     from tvsd.season import Season
     from tvsd.show import Show
+
+console = Console()
 
 
 class SearchQuery:
@@ -86,6 +90,7 @@ class SearchQuery:
         LOGGER.debug("Searching for %s", self._query)
 
         for file in dir(sources):
+            LOGGER.debug(f"Found {file}...")
             if not file.startswith("__"):
                 for cls_name, cls_obj in inspect.getmembers(
                     sys.modules[f"tvsd.sources.{file}"]
@@ -100,9 +105,14 @@ class SearchQuery:
                         LOGGER.info(f"Searching {cls_name}...")
                         query_results += cls_obj().query_from_source(self._query)
 
-        for result_index, result in enumerate(query_results):
-            print(result_index, result.title, result.source.source_name, result.note)
+        table = Table("index", "Title", "Source", "Note")
 
+        for result_index, result in enumerate(query_results):
+            table.add_row(
+                str(result_index), result.title, result.source.source_name, result.note
+            )
+
+        console.print(table)
         self._chosen_show = query_results[typer.prompt(text="请选择你下载的节目", type=int)]
 
     @property