+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## v1.0.1 (2023-07-12)
+
+### Refactor
+
+- added all sources import in __init__.py and generalized ssstv urls into source_url and relative path
+
+## v1.0.0 (2023-07-11)
+
+### Feat
+
+- **source**: Added new media source 777tv
+
+### Fix
+
+- websockets auto patch failed due to pyppeteer==1.0.2 dependencies
+
 ## v1.0.0a0 (2023-07-11)
 
 ### Fix